+++ esmerald/responses/template.py
@@ -23,7 +23,7 @@ class TemplateResponse(Response):
         background: Optional[Union["BackgroundTask", "BackgroundTasks"]] = None,
         headers: Optional[Dict[str, Any]] = None,
         cookies: Optional["ResponseCookies"] = None,
-        media_type: Union[MediaType, str] = MediaType.HTML,
+        media_type: Union[MediaType, str] = MediaType.JSON,
     ):
         if media_type == MediaType.JSON:  # we assume this is the default
             suffixes = PurePath(template_name).suffixes
@@ -38,6 +38,9 @@ class TemplateResponse(Response):
         self.template = template_engine.get_template(template_name)
         self.context = context or {}
         content = self.template.render(**context)
+        # ensure template string is not mangled
+        if isinstance(content, str):
+            content = content.encode(self.charset)
         super().__init__(
             content=content,
             status_code=status_code,